refactor(clustering): move compat code to utils
